* Various notation rendering improvements.  Still work to be done, but   things are looking up.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10408 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/editors/notation/NotationGroup.h
@@ -95,10 +95,10 @@ protected:
 private:
     struct Beam
     {                           // if a beam has a line equation y = mx + c,
-        int  gradient;          // -- then this is m*100 (i.e. a percentage)
-        int  startY;            // -- and this is c
-        bool aboveNotes;
-        bool necessary;
+	float gradient;         // -- then this is m*100 (i.e. a percentage)
+        int   startY;           // -- and this is c
+        bool  aboveNotes;
+        bool  necessary;
     };
 
     Beam calculateBeam(NotationStaff &);
